requirements.txt
Flask
Flask-SocketIO
eventlet
Pillow